+++ Control/Monad/Primitive.hs
@@ -219,11 +219,13 @@ instance PrimBase (ST s) where
   internal (ST p) = p
   {-# INLINE internal #-}
 
+-- @since 0.7.1.0
 instance PrimMonad (L.ST s) where
   type PrimState (L.ST s) = s
   primitive = L.strictToLazyST . primitive
   {-# INLINE primitive #-}
 
+-- @since 0.7.1.0
 instance PrimBase (L.ST s) where
   internal = internal . L.lazyToStrictST
   {-# INLINE internal #-}