+++ changelog.md
@@ -1,5 +1,7 @@
 ## Changes in version 0.7.1.0
 
+  * Add `shrinkMutableByteArray`.
+
   * Add `Eq` instances for `MutableByteArray` and `MutablePrimArray`.
 
 ## Changes in version 0.7.0.0